+++ src/common/utxobased/info/qtum.ts
@@ -42,7 +42,7 @@ const currencyInfo: EdgeCurrencyInfo = {
 const engineInfo: EngineInfo = {
   serverConfigs: [
     {
-      type: 'blockbook-nownode',
+      type: 'blockbook',
       uris: ['https://qtum-wusa1.edge.app']
     }
   ],